fix: lint fix

+++ packages/contracts/src/contracts/intent/FeeAdapter.sol
@@ -1,17 +1,18 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.25;
 
-import { IERC20 } from '@openzeppelin/contracts/token/ERC20/IERC20.sol';
-import { SafeERC20 } from '@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol';
-import { Address } from '@openzeppelin/contracts/utils/Address.sol';
-import { Ownable2Step, Ownable } from '@openzeppelin/contracts/access/Ownable2Step.sol';
+import {Ownable, Ownable2Step} from '@openzeppelin/contracts/access/Ownable2Step.sol';
+import {IERC20} from '@openzeppelin/contracts/token/ERC20/IERC20.sol';
+import {SafeERC20} from '@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol';
+import {Address} from '@openzeppelin/contracts/utils/Address.sol';
 
-import { TypeCasts } from 'contracts/common/TypeCasts.sol';
+import {TypeCasts} from 'contracts/common/TypeCasts.sol';
 
-import { IEverclear } from 'interfaces/common/IEverclear.sol';
-import { IFeeAdapter } from 'interfaces/intent/IFeeAdapter.sol';
-import { IEverclearSpoke } from 'interfaces/intent/IEverclearSpoke.sol';
-import { IPermit2 } from 'interfaces/common/IPermit2.sol';
+import {IEverclear} from 'interfaces/common/IEverclear.sol';
+
+import {IPermit2} from 'interfaces/common/IPermit2.sol';
+import {IEverclearSpoke} from 'interfaces/intent/IEverclearSpoke.sol';
+import {IFeeAdapter} from 'interfaces/intent/IFeeAdapter.sol';
 
 contract FeeAdapter is IFeeAdapter, Ownable2Step {
   ////////////////////
@@ -51,7 +52,9 @@ contract FeeAdapter is IFeeAdapter, Ownable2Step {
   ////////////////////
 
   /// @inheritdoc IFeeAdapter
-  function updateFeeRecipient(address _feeRecipient) external onlyOwner {
+  function updateFeeRecipient(
+    address _feeRecipient
+  ) external onlyOwner {
     _updateFeeRecipient(_feeRecipient);
   }
 
@@ -81,17 +84,8 @@ contract FeeAdapter is IFeeAdapter, Ownable2Step {
     _pullTokens(msg.sender, _inputAsset, _amount + _fee);
 
     // Create intent
-    (_intentId, _intent) = _newIntent(
-      _destinations,
-      _receiver,
-      _inputAsset,
-      _outputAsset,
-      _amount,
-      _maxFee,
-      _ttl,
-      _data,
-      _fee
-    );
+    (_intentId, _intent) =
+      _newIntent(_destinations, _receiver, _inputAsset, _outputAsset, _amount, _maxFee, _ttl, _data, _fee);
   }
 
   /// @inheritdoc IFeeAdapter
@@ -111,17 +105,8 @@ contract FeeAdapter is IFeeAdapter, Ownable2Step {
     _pullWithPermit2(_inputAsset, _amount + _fee, _permit2Params);
 
     // Call internal helper to create intent
-    (_intentId, _intent) = _newIntent(
-      _destinations,
-      _receiver,
-      _inputAsset,
-      _outputAsset,
-      _amount,
-      _maxFee,
-      _ttl,
-      _data,
-      _fee
-    );
+    (_intentId, _intent) =
+      _newIntent(_destinations, _receiver, _inputAsset, _outputAsset, _amount, _maxFee, _ttl, _data, _fee);
   }
 
   /// @inheritdoc IFeeAdapter
@@ -145,9 +130,9 @@ contract FeeAdapter is IFeeAdapter, Ownable2Step {
     // Create `_numIntents` intents with the same params and `_amount` divided
     // equally across all created intents.
     uint256 _toSend = _params.amount / _numIntents;
-    for (uint i; i < _numIntents - 1; i++) {
+    for (uint256 i; i < _numIntents - 1; i++) {
       // Create new intent
-      (bytes32 _intentId, ) = spoke.newIntent(
+      (bytes32 _intentId,) = spoke.newIntent(
         _params.destinations,
         _params.receiver,
         _params.inputAsset,
@@ -161,7 +146,7 @@ contract FeeAdapter is IFeeAdapter, Ownable2Step {
     }
 
     // Create a final intent here with the remainder of balance
-    (bytes32 _intentId, ) = spoke.newIntent(
+    (bytes32 _intentId,) = spoke.newIntent(
       _params.destinations,
       _params.receiver,
       _params.inputAsset,
@@ -195,7 +180,7 @@ contract FeeAdapter is IFeeAdapter, Ownable2Step {
 
       // Get the sum of the order amounts
       uint256 _orderSum;
-      for (uint i; i < _numIntents; i++) {
+      for (uint256 i; i < _numIntents; i++) {
         _orderSum += _params[i].amount;
         if (_params[i].inputAsset != _asset) {
           revert MultipleOrderAssets();
@@ -214,9 +199,9 @@ contract FeeAdapter is IFeeAdapter, Ownable2Step {
 
     // Initialising array length
     _intentIds = new bytes32[](_numIntents);
-    for (uint i; i < _numIntents; i++) {
+    for (uint256 i; i < _numIntents; i++) {
       // Create new intent
-      (bytes32 _intentId, ) = spoke.newIntent(
+      (bytes32 _intentId,) = spoke.newIntent(
         _params[i].destinations,
         _params[i].receiver,
         _params[i].inputAsset,
@@ -272,16 +257,8 @@ contract FeeAdapter is IFeeAdapter, Ownable2Step {
     _approveSpokeIfNeeded(_inputAsset, _amount);
 
     // Create new intent
-    (_intentId, _intent) = spoke.newIntent(
-      _destinations,
-      _receiver,
-      _inputAsset,
-      _outputAsset,
-      _amount,
-      _maxFee,
-      _ttl,
-      _data
-    );
+    (_intentId, _intent) =
+      spoke.newIntent(_destinations, _receiver, _inputAsset, _outputAsset, _amount, _maxFee, _ttl, _data);
 
     // Emit event
     emit IntentWithFeesAdded(_intentId, msg.sender.toBytes32(), _fee, msg.value);
@@ -292,7 +269,9 @@ contract FeeAdapter is IFeeAdapter, Ownable2Step {
    * @notice Updates the fee recipient
    * @param _feeRecipient New recipient
    */
-  function _updateFeeRecipient(address _feeRecipient) internal {
+  function _updateFeeRecipient(
+    address _feeRecipient
+  ) internal {
     emit FeeRecipientUpdated(_feeRecipient, feeRecipient);
     feeRecipient = _feeRecipient;
   }
@@ -324,7 +303,7 @@ contract FeeAdapter is IFeeAdapter, Ownable2Step {
     // Checking if the strategy is default or not
     address spender;
     IEverclear.Strategy _strategy = spoke.strategies(_asset);
-    if(_strategy == IEverclear.Strategy.DEFAULT) spender = address(spoke);
+    if (_strategy == IEverclear.Strategy.DEFAULT) spender = address(spoke);
     else spender = xerc20Module;
 
     // Approve the spoke contract if needed
@@ -358,11 +337,11 @@ contract FeeAdapter is IFeeAdapter, Ownable2Step {
     // Transfer from caller using permit2
     PERMIT2.permitTransferFrom(
       IPermit2.PermitTransferFrom({
-        permitted: IPermit2.TokenPermissions({ token: IERC20(_asset), amount: _amount }),
+        permitted: IPermit2.TokenPermissions({token: IERC20(_asset), amount: _amount}),
         nonce: _permit2Params.nonce,
         deadline: _permit2Params.deadline
       }),
-      IPermit2.SignatureTransferDetails({ to: address(this), requestedAmount: _amount }),
+      IPermit2.SignatureTransferDetails({to: address(this), requestedAmount: _amount}),
       msg.sender,
       _permit2Params.signature
     );
